+++ py_ballisticcalc.exts/py_ballisticcalc_exts/rk4_engine.pyx
@@ -17,6 +17,6 @@ __all__ = [
 cdef class CythonizedRK4IntegrationEngine(CythonizedBaseIntegrationEngine):
     """Cythonized RK4 (Runge-Kutta 4th order) integration engine for ballistic calculations."""
 
-    def __cinit__(self, object _config):
+    def __cinit__(self, object config):
         self._DEFAULT_TIME_STEP = 0.0025
         self._this.integrate_func = BCLIBC_IntegrateCallable(BCLIBC_integrateRK4)